+++ CHANGELOG.md
@@ -1,12 +1,21 @@
 # PostgreSQL driver for Yii Database Change Log
 
-## 1.1.1 under development
+## 1.2.1 under development
 
-- Enh #301: Refactor `JsonExpressionBuilder` (@Tigrov)
+- Bug #316, #6: Support table view constraints (@Tigrov)
+- Enh #324: Change property `Schema::$typeMap` to constant `Schema::TYPE_MAP` (@Tigrov)
+- Enh #303: Support composite types (@Tigrov)
+
+## 1.2.0 November 12, 2023
+
+- Chg #319: Remove use of abstract type `SchemaInterface::TYPE_JSONB` (@Tigrov)
 - Enh #300: Refactor `ArrayExpressionBuilder` (@Tigrov)
+- Enh #301: Refactor `JsonExpressionBuilder` (@Tigrov)
 - Enh #302: Refactor `ColumnSchema` (@Tigrov)
+- Enh #321: Move methods from `Command` to `AbstractPdoCommand` class (@Tigrov)
 - Bug #302: Fix incorrect convert string value for BIT type (@Tigrov)
 - Bug #309: Fix retrieving sequence name from default value (@Tigrov)
+- Bug #313: Refactor `DMLQueryBuilder`, related with yiisoft/db#746 (@Tigrov)
 
 ## 1.1.0 July 24, 2023
 
@@ -18,7 +27,7 @@
 - Enh #294: Refactoring of `Schema::normalizeDefaultValue()` method (@Tigrov)
 - Bug #287: Fix `bit` type (@Tigrov)
 - Bug #295: Fix multiline and single quote in default string value, add support for PostgreSQL 9.4 parentheses around negative numeric default values (@Tigrov)
-- Bug #296: Prevent posible issues with array default values `('{one,two}'::text[])::varchar[]`, remove `ArrayParser::parseString()` (@Tigrov)
+- Bug #296: Prevent possible issues with array default values `('{one,two}'::text[])::varchar[]`, remove `ArrayParser::parseString()` (@Tigrov)
 
 ## 1.0.0 April 12, 2023